refactor(client): consolidate client navigation logic
Refactor the client navigation structure by creating a dedicated
ClientNavGraph.kt file. This change improves code organization
and maintainability by encapsulating all client-related navigation
in one place, making it easier to manage and extend in the future.
